Fix FindOperatorPairs did not find expected number of op_ methods
- https://github.com/mono/Embeddinator-4000/issues/207
- Multiple operators of the same type but different arguments would fail unnecessarily
- Move overloads tests to new file due to length.
